+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Version](http://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 ### Added
 * Bugfix: Elasticsearch magento cli installation parameters not available for Magento 2.3.X
+* Feature: SampleData optionally installable during the project creation 
 ### Changed
 ### Removed